Use the Backend to check for dask dataframes, instead of a hard check.
