+++ ludwig/constants.py
@@ -172,3 +172,5 @@ HYPEROPT_WARNING = (
 CONTINUE_PROMPT = "Do you want to continue? "
 
 DEFAULT_AUDIO_TENSOR_LENGTH = 70000
+
+DASK_MODULE_NAME = "dask.dataframe"
